flask-sqlserver-app/requirements.txt
Flask
flask-sqlalchemy
pyodbc
flask-migrate
flask-cors
python-dotenv